refactor: useEffect의 콜백함수 분리
